Enabled UnusedPrivateMethod pmd rule
