+++ collect_app/src/androidTest/java/org/odk/collect/android/support/pages/QRCodePage.java
@@ -1,18 +1,12 @@
 package org.odk.collect.android.support.pages;
 
 import android.graphics.Bitmap;
-import android.graphics.drawable.BitmapDrawable;
-import android.graphics.drawable.Drawable;
-import android.view.View;
-import android.widget.ImageView;
 
 import androidx.test.espresso.Espresso;
-import androidx.test.espresso.matcher.BoundedMatcher;
 
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
 import org.odk.collect.android.R;
 import org.odk.collect.android.support.ActivityHelpers;
+import org.odk.collect.android.support.matchers.DrawableMatcher;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
@@ -58,33 +52,4 @@ public class QRCodePage extends Page<QRCodePage> {
 
         return this;
     }
-
-    // Matcher class to match the contents of a ImageView and compare with a bitmap
-    private static class DrawableMatcher {
-        private static Matcher<View> withBitmap(Bitmap match) {
-            return new BoundedMatcher<View, ImageView>(ImageView.class) {
-                @Override
-                public void describeTo(Description description) {
-                    description.appendText("bitmaps did not match");
-                }
-
-                @Override
-                protected boolean matchesSafely(ImageView imageView) {
-                    Drawable drawable = imageView.getDrawable();
-                    if (drawable == null && match == null) {
-                        return true;
-                    } else if (drawable != null && match == null) {
-                        return false;
-                    } else if (drawable == null && match != null) {
-                        return false;
-                    }
-
-                    Bitmap actual = ((BitmapDrawable) drawable).getBitmap();
-
-                    return actual.sameAs(match);
-                }
-            };
-        }
-    }
-
 }
